feat: agregacion de logs en tiempo real

+++ src/main/java/com/safiraenergia/mercadospot/dto/factura/FacturaDTO.java
@@ -16,9 +16,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class FacturaDTO {
     private Long folio;
-    private int montoNeto;
-    private int montoBruto;
-    private int montoTotal;
+    private Double montoNeto;
+    private Double montoBruto;
+    private Double montoTotal;
     private Date fechaEmision;
     private Date fechaPago;
     private Date periodo;
